[API] Refactor the fn list_all/1 to all/1
I've done this change to adequate it that what I want to do.

The idea now is something like this:
```elixir
base_query()
|> filter_query()
|> order_query()
|> Repo.all
```

With that I believe that I can accomplishiment the requirements related to fetch Players!

I used this referenc to get this result https://elixirschool.com/blog/ecto-query-composition/